+++ review/lib/github/github_test.go
@@ -1,0 +1,11 @@
+package github
+
+import (
+	"testing"
+)
+
+func TestGitHubPackageExists(t *testing.T) {
+	// This is a simple test to ensure the package compiles
+	// Real tests would require GitHub API mocking which is beyond scope
+	t.Log("GitHub API package is properly structured")
+}
